Move editing state up into TodoList so we don't lose it when indenting
Without this, given that we want to keep a proper DOM hierarchy that matches our Todo tree,
we lose state every time we indent, as React can't track state when reparenting.

With this change, the state is in the list, tracked against todo id's, so doesn't get lost.
The list then just uses it to select between the two different Todo item presentation components
available as appropriate.