Fix bug with child placement when unindenting
Previously a child would always end up as the last child of the parent it moved to. That's fine most
of the time, but if you're the first child of the first child of the root node, and there are ten
other children of the root node, when you pop up you want to appear as the second child of the root node
not way down the list. This makes sure we bump you up into the right slot.