Fix bug where we lose state of child todo items when editing/unediting the parent
This happens because we're effectively reparenting them by swapping out the Editable/normal todo items.
We now keep the state outside the todo item, but param's it in, so that editing state is help independently
of the nodes that we're moving around completely

This also drastically improves the handling of event bubbling in their hierarchy structure, by putting this
onus on the 'start editing' handler to check exactly what was clicked on. Much tidier.